add telegram executor

+++ src/main/java/weather/WeatherRequest.java
@@ -74,7 +74,7 @@ public class WeatherRequest {
 
     }
 
-    public String getForecastByGeoCoord(Double latitude, Double longitude){
+    public String getForecastByGeoCoord(Float latitude, Float longitude){
 
         //https://samples.openweathermap.org/data/2.5/forecast?lat=35&lon=139&appid=b6907d289e10d714a6e88b30761fae22
